+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Add a Stimulus controller to disable expired links.
 * Add `temporary_url_autoexpire` Twig function.
+* Add expiration in the query string of the resulting URL.
 
 ## 1.2.1